backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.2
feedparser==6.0.11
apscheduler==3.10.4
pydantic-settings==2.5.2
